python-projets/projet-cloud-backup/requirements.txt
azure-core==1.39.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.12
isodate==0.7.2
msal==1.36.0
msal-extensions==1.3.1
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
